Remove with_replies option
Twitter have decided that replies ought not to be publicly available in their web ui without logging in

I think it is a troubling move towards the privatized "walled garden" mentality of FB et al.

The consequence for TwitRSS.me is that it will no longer be able to support the with_replies feature. Pity. That was useful.

This commit removes the feature from the UI. Its still there in the API. Just won't return ant results.

References: #33