Finish implementation of using mobile twitter site
as basis of parsing instead of normal twitter.

Some limitations:

* No images are available. The images aren't available on the feed so it
isn't possible
* Dates are somewhat inaccurate. The year month and day are most likely
correct (test cases where not correct are most welcome). However the
actual dates are not in the feed, so again, it isn't possible to be more
accurate than the day.

I should probably refactor this work so that the parsing is in modules,
but want to see if it works first of all.